Use o65 file for kernel
Well, this was about a day of work but honestly I don't know if I want
to keep going with it. Waiting minutes for a microcomputer to boot up is
not acceptable, and the kernel will just be too large to do all of the
o65 parsing. Additionally, the kernel doesn't need to be relocated so
the benefit is not there.

I am still commiting this so that it is saved somewhere, but really I
think having the kernel as a binary is a better option, and we should
leave o65 for the user programs.