Rewrite o65 in assembly again
I wrote a test program to try and test it, turns out when they say a
location is callee saved, that means the caller should probably save it.
Instead of doing that, I switched to using callee saved registers, then
I can just save all 6 of them at the start and restore them at the end,
instead of saving and restoring the temporary values all the time.

This does make the code a little bit larger but it now makes it exactly
half the size of the C version, which I think is fun.