Fix init, readblock
The problem with init seemed to be that were were not checking to see if
the card was ready after ACMD41. Now, we check the power up status and
issue more ACMD41s if neccesary. We also get rid of the get rca function
and return the RCA in init. This means that after init, the card is in
the standby state in data transfer mode, which makes more sense.

Readblock was not accessing the parameters off of the stack correctly,
so that is fixed now and the write address is read.

Overall, it is amazing how slow the C code was before.