throw some stuff on.
This will have to be changed, some things like the VGA connector will
not work because they will interfere with the cpu adapter.

It may be worth thinking about making these smaller and not including
the ethernet for example, since that will not come for a while.

The only things we really need to start are the SD card and serial port.