Update sdram controller with new Efinity version
For some reason the old one did not want to open in the new version.
Even though the version of the IP is the same, something about it being
made with the old Efinity version made it mad.

I just deleted it and made it again with the same settings and now it
lets me open and configure it.