Change sd clock phase, remove check for flags
Previously it would only update the flags if the sd clock is in a
certain phase. In the previous system the SD card clock ran at the same
speed as the cpu clock, but now the sd card clock is half the cpu clock
so half the time the read flag will not be reset, depending on how the
phases aligned.

Now there is no check so it should always work.